Redraw the 3-D scanner the way WC2 does
The scanner grid took its colour from the text and palette globals where
WC2 uses two literals, 0x5b for the forward view and 0x25 for the target
camera, so both grids came out in the wrong colour.  The target marker
also cached the target index in the loop variable left over from the
contact sweep.

vid_equiv and check_stranded each carried an extra guard - the canned
scene mode and the train-sim state - that WC2 does not test, so both went
quiet in situations where the retail game still speaks.

+++ include/wc1funcs.h
@@ -804,7 +804,7 @@ void UpdateTargetCameraTracking(void);                                 /* 0x4608
 void ClearSavedTorpedoTargetLock(void);                                /* 0x460BAF */
 void UpdateSavedTorpedoTargetLock(void);                               /* 0x460BCC */
 void SetTargetCameraEyePosition(short cameraMode);                     /* 0x460D3E */
-void DrawTargetCameraAttitudeIndicators(unsigned int colour);          /* 0x460DFE */
+void DrawTargetCameraAttitudeIndicators(unsigned short colour);       /* 0x460DFE */
 void RestoreTargetCameraAttitudeIndicators(void);                      /* 0x4610BC */
 void UpdateTargetCameraCockpitHook(void);                              /* 0x461456 */
 void FinalizeTargetCameraViewHook(void);                               /* 0x461466 */
